+++ src/products/mutations.ts
@@ -58,6 +58,10 @@ import {
   ProductVariantReorderVariables
 } from "./types/ProductVariantReorder";
 import {
+  ProductVariantSetDefault,
+  ProductVariantSetDefaultVariables
+} from "./types/ProductVariantSetDefault";
+import {
   SimpleProductUpdate,
   SimpleProductUpdateVariables
 } from "./types/SimpleProductUpdate";
@@ -133,6 +137,26 @@ export const useProductImagesReorder = makeMutation<
   ProductImageReorder,
   ProductImageReorderVariables
 >(productImagesReorder);
+
+const productVariantSetDefault = gql`
+  ${productErrorFragment}
+  ${productFragmentDetails}
+  mutation ProductVariantSetDefault($productId: ID!, $variantId: ID!) {
+    productVariantSetDefault(productId: $productId, variantId: $variantId) {
+      errors: productErrors {
+        ...ProductErrorFragment
+      }
+      product {
+        ...Product
+      }
+    }
+  }
+`;
+
+export const useProductVariantSetDefaultMutation = makeMutation<
+  ProductVariantSetDefault,
+  ProductVariantSetDefaultVariables
+>(productVariantSetDefault);
 
 export const productUpdateMutation = gql`
   ${productErrorFragment}